# backend/public/api-docs.yaml
openapi: 3.0.0
info:
  title: 'AlleyNote API'
  description: "\nAlleyNote 公布欄系統 API 文件\n\n這是一個現代化的公布欄系統 API，提供完整的貼文管理、使用者認證、檔案上傳等功能。\n\n## 認證方式\n- **JWT Bearer Token**: 使用 JWT Token 進行 API 認證\n- **Session Auth**: 傳統 Session 認證 (向下相容)\n\n## 主要功能\n- 📝 **貼文管理**: 建立、編輯、刪除、查看貼文\n- 👤 **使用者認證**: 登入、登出、註冊、密碼管理\n- 📎 **檔案上傳**: 支援附件上傳和管理\n- 🔐 **權限控制**: 基於角色的存取控制\n\n## API 版本\n當前版本: v1.0.0\n支援的回應格式: JSON\n字元編碼: UTF-8\n\n## 錯誤處理\n所有錯誤回應都遵循統一格式：\n```json\n{\n  \"success\": false,\n  \"error\": \"錯誤訊息\",\n  \"errors\": {\n    \"field\": [\"具體錯誤說明\"]\n  }\n}\n```\n\n## 聯絡資訊\n如有問題或建議，請聯繫開發團隊。\n    "
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
    description: 開發環境伺服器
  -
    url: 'https://api.alleynote.example.com'
    description: 正式環境伺服器
paths:
...
  '/api/posts/{id}':
    get:
      tags:
        - posts
      summary: 取得單一貼文
      description: '根據 ID 取得貼文詳細資訊，並記錄瀏覽次數'
      operationId: getPostById
      parameters:
        -
          name: id
          in: path
          description: '貼文 ID'
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: 成功取得貼文詳細資訊
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/Post' }
                type: object
        '404':
          description: 貼文不存在
          content:
            application/json:
              schema:
                $ref: '#/components/responses/NotFound'
    put:
      tags:
        - posts
      summary: 更新貼文
      description: '更新指定 ID 的貼文，需要 CSRF Token 驗證'
      operationId: updatePost
      parameters:
        -
          name: id
          in: path
          description: '貼文 ID'
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        description: 更新的貼文資料
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostRequest'
      responses:
        '200':
          description: 貼文更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 輸入資料驗證失敗
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: 未授權存取
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Unauthorized'
        '403':
          description: 'CSRF Token 驗證失敗'
          content:
            application/json:
              schema:
                properties:
                  error: { type: string, example: 'CSRF token verification failed' }
                type: object
        '404':
          description: 貼文不存在
          content:
            application/json:
              schema:
                $ref: '#/components/responses/NotFound'
      security:
        -
          csrfToken: []
    delete:
      tags:
        - posts
      summary: 刪除貼文
      description: '刪除指定 ID 的貼文，需要 CSRF Token 驗證'
      operationId: deletePost
      parameters:
        -
          name: id
          in: path
          description: '貼文 ID'
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '204':
          description: 貼文刪除成功
        '401':
          description: 未授權存取
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Unauthorized'
        '403':
          description: 'CSRF Token 驗證失敗'
          content:
            application/json:
              schema:
                properties:
                  error: { type: string, example: 'CSRF token verification failed' }
                type: object
        '404':
          description: 貼文不存在
          content:
            application/json:
              schema:
                $ref: '#/components/responses/NotFound'
      security:
        -
          csrfToken: []
...
  /api/tags:
    get:
      tags:
        - Tags
      summary: 'Get list of tags'
      description: "取得標籤列表.\n\nGET /api/tags"
      operationId: getTags
      parameters:
        -
          name: page
          in: query
          schema:
            type: integer
            default: 1
        -
          name: per_page
          in: query
          schema:
            type: integer
            default: 20
        -
          name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 'Tags retrieved successfully'
        '500':
          description: 'Internal server error'
    post:
      tags:
        - Tags
      summary: 'Create a new tag'
      description: "建立標籤.\n\nPOST /api/tags"
      operationId: createTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
              properties:
                name:
                  type: string
                  maxLength: 50
                slug:
                  type: string
                description:
                  type: string
                color:
                  type: string
              type: object
      responses:
        '201':
          description: 'Tag created successfully'
        '422':
          description: 'Validation failed'
  '/api/tags/{id}':
    get:
      tags:
        - Tags
      summary: 'Get a single tag'
      description: "取得單一標籤.\n\nGET /api/tags/{id}"
      operationId: getTag
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Tag retrieved successfully'
        '404':
          description: 'Tag not found'
    put:
      tags:
        - Tags
      summary: 'Update a tag'
      description: "更新標籤.\n\nPUT /api/tags/{id}"
      operationId: updateTag
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  maxLength: 50
                slug:
                  type: string
                description:
                  type: string
                color:
                  type: string
              type: object
      responses:
        '200':
          description: 'Tag updated successfully'
        '404':
          description: 'Tag not found'
        '422':
          description: 'Validation failed'
    delete:
      tags:
        - Tags
      summary: 'Delete a tag'
      description: "刪除標籤.\n\nDELETE /api/tags/{id}"
      operationId: deleteTag
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Tag deleted successfully'
        '404':
          description: 'Tag not found'
...
components:
  schemas:
...
    UpdatePostRequest:
      title: 更新貼文請求
      description: 更新既有貼文的請求資料，所有欄位皆為可選，僅更新提供的欄位
      properties:
        title:
          description: 更新貼文標題
          type: string
          maxLength: 255
          minLength: 1
          example: 【更新】系統維護延期通知
        content:
          description: 更新貼文內容
          type: string
          maxLength: 10000
          minLength: 1
          example: "**重要更新**\n\n原定於 2025年1月20日 的系統維護將延期至 **2025年1月25日**。\n\n詳細資訊請參考最新公告。"
        category:
          description: 更新貼文分類
          type: string
          enum:
            - general
            - announcement
            - urgent
            - notice
          example: urgent
        status:
          description: 更新貼文狀態
          type: string
          enum:
            - draft
            - published
            - archived
          example: published
        priority:
          description: 更新優先級
          type: string
          enum:
            - 1
            - 2
            - 3
            - 4
            - 5
          example: urgent
        expires_at:
          description: 更新過期時間
          type: string
          format: date-time
          example: '2025-02-28T23:59:59+08:00'
          nullable: true
        tags:
          description: 更新標籤列表
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 10
          example:
            - 維護延期
            - 重要更新
            - 緊急通知
        is_pinned:
          description: 更新置頂狀態
          type: boolean
          example: true
        allow_comments:
          description: 更新留言允許狀態
          type: boolean
          example: true
      type: object
      example:
        title: 【更新】系統維護延期通知
        content: "**重要更新**\n\n原定於 2025年1月20日 的系統維護將延期..."
        category: urgent
        status: published
        priority: urgent
        expires_at: '2025-02-28T23:59:59+08:00'
        tags:
          - 維護延期
          - 重要更新
          - 緊急通知
        is_pinned: true
        allow_comments: true
    ApiResponse:
      title: 'API 回應'
      description: '標準 API 回應格式，用於包裝所有 API 的成功回應'
      required:
        - success
      properties:
        success:
          description: 操作是否成功
          type: boolean
          example: true
        message:
          description: 操作結果訊息
          type: string
          example: 貼文建立成功
        data:
          description: '回應資料內容，根據不同 API 會有不同結構'
          oneOf:
            -
              $ref: '#/components/schemas/Post'
            -
              type: array
              items:
                $ref: '#/components/schemas/Post'
            -
              description: 其他類型的資料
              type: object
        meta:
          description: 額外的元資料
          properties:
            timestamp:
              type: string
              format: date-time
              example: '2025-01-15T10:30:00+08:00'
            request_id:
              type: string
              example: req_550e8400e29b41d4a716446655440000
            version:
              type: string
              example: 1.0.0
          type: object
      type: object
      example:
        success: true
        message: 貼文建立成功
        data:
          id: 123
          title: 新的貼文標題
          content: 貼文內容...
          status: published
          created_at: '2025-01-15T10:30:00+08:00'
        meta:
          timestamp: '2025-01-15T10:30:00+08:00'
          request_id: req_550e8400e29b41d4a716446655440000
          version: 1.0.0
...
    Post:
      title: 貼文
      description: 貼文資料結構
      required:
        - id
        - title
        - content
        - status
        - created_at
      properties:
        id:
          description: 貼文ID
          type: integer
          example: 1
        title:
          description: 貼文標題
          type: string
          maxLength: 255
          example: 重要公告
        content:
          description: 貼文內容
          type: string
          example: 這是一則重要公告內容
        category:
          description: 貼文分類
          type: string
          enum:
            - general
            - announcement
            - urgent
            - notice
          example: announcement
        status:
          description: 貼文狀態
          type: string
          enum:
            - draft
            - published
            - archived
          example: published
        priority:
          description: 優先級
          type: string
          enum:
            - 1
            - 2
            - 3
            - 4
            - 5
          example: normal
        author_id:
          description: 作者ID
          type: integer
          example: 1
        created_at:
          description: 建立時間
          type: string
          format: date-time
          example: '2025-01-15T10:30:00Z'
        updated_at:
          description: 更新時間
          type: string
          format: date-time
          example: '2025-01-15T11:00:00Z'
        expires_at:
          description: 過期時間
          type: string
          format: date-time
          example: '2025-12-31T23:59:59Z'
          nullable: true
      type: object
  responses:
    Unauthorized:
      description: 未授權存取
      content:
        application/json:
          schema:
            properties:
              success:
                description: "OpenAPI 規格配置.\n\n此檔案定義 AlleyNote API 的基本 OpenAPI 規格資訊"
                type: boolean
                example: false
              error:
                type: string
                example: 未授權存取
            type: object
...
    NotFound:
      description: 資源不存在
      content:
        application/json:
          schema:
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: 資源不存在
            type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      name: Authorization
      in: header
      bearerFormat: JWT
      scheme: bearer
    sessionAuth:
      type: apiKey
      name: PHPSESSID
      in: cookie
...